Move test Selector to /test
